Refactored Project back to Jigsaw
I know, I know..."Make up your mind!" Well, here we are...

`ServiceLoaders` were not working at all in the standard Jar Java project. I do mean "at all"! Where I was able to get the main top-level menus to load dynamically in the old Jigsaw project, not even those would load in the standard Java project. Now, I have not tested it again in this project yet, as I just wanted to get this refactoring up to the repository. So, there it is.

I would like to point out that resources ***are*** working! Since placing the `opens <package> to <module>` statements did not work, I tried changing the module declaration to `open module <module>`. Lo and behold! That worked! the AppFramework is now able to process components and load resources. This is a major step!

This is the project that we are going to work with from here on out. Big refactorings and project changes are now officially in the past. We go from here, or we do not go...

-SC