Dynamic Menus and Menu Items are Working!
I have gotten all of the kinks worked out of the dynamic menu system...I believe. I still haven't tested having another module provide a menu or menu item, but it is currently working for I have. I will continue working on this.

Also, due to changing the modules to Open Modules, all of the resources are working as expected. Not only is text being pulled from the resource files, but so are icons. Still need to check colors and fonts, but if the icons are working, I am willing to be that colors and fonts are too.

It is important to note that I have set the interfaces up as standard interfaces in the API. I did this because I was not able to get the annotations to work, at least not consistently. Therefore, I decided to drop back to standard interfaces and leave annotations for a future change, if we decide to. If we do that though, we will need to decorate the standard interfaces with `@Deprecated` and leave them in the API until we can get them all changed out to annotations. Just a thought.

-SC