Testing from Other Modules
I have added an `OptionsPanelProvider` and a `MenuProvider` to the `NTA.Accounting` module to test that the dynamic generation is truly working. Ran the project and the Manage menu and Accounting Defaults options panel both displayed as expected.

I also created the About screen and provided a manner by which modules could provide about information for themselves. Instead of having the modules create a subclass of `JPanel` and implementing an API to mark it as being for the About box (like with the `OptionsPanelProvider`), I made the standard `JPanel` for each module that will be dynamically populated with the information provided by the modules `AboutPanelProvider` class. The `AboutPanelProvider` allows the module to provide the following information:

- Title: Typically, this would be the name of the module, such as "NTA.Accounting", or a one- or two-word description of the feature, such as "Accounting" or "Load Tracking". This title will be set as the tab title for the module's about panel.
- Contributor(s): This is a `String[]` of the contributors to the module.
- Vendor: This is the name of the individual or company that provided the module.
- Copyright: This is the copyright line for the module, such as "Copyright© 2006-2021 ${CompanyOrIndividual}. All rights under copyright reserved."
- Logo: This is so the module may provide a 48x48 logo icon for their module.
- License URL: this is so the module may provide a URL to a website containing the full license under which the module is release. This URL is placed into a link label, so the user can view the license if they choose do so.

The reason I did the `AboutPanelProvider` in this manner, instead of just letting modules extend a `JPanel` and implement the `AboutPanelProvider` is to keep the appearance of each module's about panel consistent:


I have also worked over the code quite a bit and found/fixed some issues. I know that we will find other issues with this dynamic loading code that will need to be fixed as we find it. However, I have placed numerous log messages in this code to allow us to track what it is doing internally at run-time, so tracking down the issues is a pretty simple process. Fixing them, on the other hand, is a completely different story...

-SC